requirements.txt
dimod==0.12.20
dwave-neal==0.6.0
dwave-samplers==1.6.0
networkx==3.5
numpy==2.3.2
pandas==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
